+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [UUID to Name history](https://wiki.vg/Mojang_API#UUID_-.3E_Name_history) endpoint.
 - [Playernames -> UUIDs](https://wiki.vg/Mojang_API#Playernames_-.3E_UUIDs) endpoint.
 - [Change Skin](https://wiki.vg/Mojang_API#Change_Skin) endpoint.
+- [Reset Skin](https://wiki.vg/Mojang_API#Reset_Skin) endpoint.
 
 ### Changed
 - The constructor no longer has arguments.